+++ spec/fixtures/multi-swift/Source/MultiSwift.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+
+public protocol Animal {
+    static var favoriteFood: String { get }
+}
+
+public struct Cat: Animal {
+    public static var favoriteFood: String {
+        return "Kibbles"
+    }
+}